Strategy pattern where each option type has its own rendering plan
